fix: rename from solutionformat -> solution
